requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
charset-normalizer==2.1.1
deeppyer==1.0.1
ffmpeg-python==0.2.0
frozenlist==1.3.1
future==0.18.2
idna==3.4
imageio-ffmpeg==0.4.7
multidict==6.0.2
numpy==1.23.3
opencv-python==4.6.0.66
Pillow==9.2.0
py-cord==2.1.3
pycodestyle==2.9.1
python-dotenv==0.21.0
toml==0.10.2
yarl==1.8.1